RD-6386: Extracted deployOn dropdown as an external component
